requirements.txt
# Flask Framework
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
pymongo==4.6.0

# Database Drivers
PyMySQL==1.1.0
# OR use: mysqlclient==2.2.0

# Cryptography Libraries
cryptography==41.0.7
pycryptodome==3.19.0

# Excel/Data Processing
openpyxl==3.1.2
#pandas==2.1.4

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1

# Optional: For production deployment
# waitress==2.1.2
# gunicorn==21.2.0

# Optional: For charts/visualization
# matplotlib==3.8.2
# Pillow==10.1.0
